Refactor out LinearAdapter to commons
