+++ examples/asr/asr_adapters/conf/asr_adaptation.yaml
@@ -8,7 +8,7 @@ model:
 
   adapter:
     adapter_name: ???  # Name of the adapter, used by the script
-    _target_: nemo.collections.asr.parts.submodules.adapter_modules.LinearAdapter
+    _target_: nemo.collections.common.parts.adapter_modules.LinearAdapter
     in_features: ???  # User must provide the output dimension of the layers of the model
     dim: 32  # can be any hidden dimension as chosen by user, but small values are preferred to reduce param count
     activation: swish
