+++ nemo/collections/common/parts/utils.py
@@ -16,7 +16,18 @@ import math
 import os
 from typing import List
 
+import torch.nn as nn
+
 __all__ = ['if_exist', '_compute_softmax']
+
+activation_registry = {
+    "hardtanh": nn.Hardtanh,
+    "relu": nn.ReLU,
+    "selu": nn.SELU,
+    "swish": nn.SiLU,
+    "silu": nn.SiLU,
+    "gelu": nn.GELU,
+}
 
 
 def if_exist(outfold: str, files: List[str]):
